--- hunmin-api/src/main/kotlin/com/api/ApiApplication.kt
@@ -1,16 +1,0 @@
-package com.api
-
-import com.application.config.ApplicationConfig
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-import org.springframework.boot.runApplication
-import org.springframework.context.annotation.Import
-
-@ConfigurationPropertiesScan
-@Import(ApplicationConfig::class)
-@SpringBootApplication
-class ApiApplication
-
-fun main(args: Array<String>) {
-    runApplication<ApiApplication>(*args)
-}
