--- hunmin-api/src/main/kotlin/com/api/HealthController.kt
@@ -1,16 +1,0 @@
-package com.api
-
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class HealthController {
-
-    @GetMapping("/health")
-    fun health() =
-        "up"
-
-    @GetMapping("/exception")
-    fun exception() =
-        IllegalArgumentException("예외 발생!")
-}
