--- hunmin-api/src/main/kotlin/com/api/auth/in/response/AuthResponse.kt
@@ -1,11 +1,0 @@
-package com.api.auth.`in`.response
-
-import io.swagger.v3.oas.annotations.media.Schema
-
-data class AuthResponse(
-    @Schema(
-        description = "Access token",
-        example = "tokeninfo123123...asdvio3",
-    )
-    val accessToken: String,
-)
