--- hunmin-api/src/main/kotlin/com/api/global/config/multimodule/ModuleConfig.kt
@@ -1,16 +1,0 @@
-package com.api.global.config.multimodule
-
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-@ComponentScan(
-    basePackages = [
-        "com.common",
-        "com.domain",
-        "com.application",
-        "com.persistence",
-        "com.adapter",
-    ],
-)
-class ModuleConfig
