--- hunmin-api/src/main/kotlin/com/api/global/config/multimodule/PropertySourceScanConfig.kt
@@ -1,17 +1,0 @@
-package com.api.global.config.multimodule
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.PropertySource
-
-@PropertySource(
-    value = [
-        "classpath:application-common-\${spring.profiles.active}.yml",
-//        "classpath:application-domain-\${spring.profiles.active}.yml",
-        "classpath:application-persistence-\${spring.profiles.active}.yml",
-//        "classpath:application-application-\${spring.profiles.active}.yml",
-//        "classpath:application-adapter-\${spring.profiles.active}.yml",
-    ],
-    factory = YamlPropertySourceFactory::class,
-)
-@Configuration
-class PropertySourceScanConfig
