+++ hunmin-bootstrap/src/main/kotlin/com/bootstrap/ApiApplication.kt
@@ -1,0 +1,16 @@
+package com.bootstrap
+
+import com.application.config.ApplicationConfig
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+import org.springframework.boot.runApplication
+import org.springframework.context.annotation.Import
+
+@ConfigurationPropertiesScan
+@Import(ApplicationConfig::class)
+@SpringBootApplication
+class ApiApplication
+
+fun main(args: Array<String>) {
+    runApplication<ApiApplication>(*args)
+}
