+++ hunmin-bootstrap/src/main/kotlin/com/bootstrap/auth/in/response/AuthResponse.kt
@@ -1,0 +1,11 @@
+package com.bootstrap.auth.`in`.response
+
+import io.swagger.v3.oas.annotations.media.Schema
+
+data class AuthResponse(
+    @Schema(
+        description = "Access token",
+        example = "tokeninfo123123...asdvio3",
+    )
+    val accessToken: String,
+)
