+++ hunmin-bootstrap/src/main/kotlin/com/bootstrap/global/config/multimodule/PropertySourceScanConfig.kt
@@ -1,0 +1,17 @@
+package com.bootstrap.global.config.multimodule
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.PropertySource
+
+@PropertySource(
+    value = [
+        "classpath:application-common-\${spring.profiles.active}.yml",
+//        "classpath:application-domain-\${spring.profiles.active}.yml",
+        "classpath:application-persistence-\${spring.profiles.active}.yml",
+//        "classpath:application-application-\${spring.profiles.active}.yml",
+//        "classpath:application-adapter-\${spring.profiles.active}.yml",
+    ],
+    factory = YamlPropertySourceFactory::class,
+)
+@Configuration
+class PropertySourceScanConfig
