+++ hunmin-bootstrap/src/main/kotlin/com/bootstrap/global/config/multimodule/YamlPropertySourceFactory.kt
@@ -1,0 +1,19 @@
+package com.bootstrap.global.config.multimodule
+
+import org.springframework.beans.factory.config.YamlPropertiesFactoryBean
+import org.springframework.core.env.PropertiesPropertySource
+import org.springframework.core.io.support.DefaultPropertySourceFactory
+import org.springframework.core.io.support.EncodedResource
+
+class YamlPropertySourceFactory : DefaultPropertySourceFactory() {
+    override fun createPropertySource(
+        name: String?,
+        resource: EncodedResource,
+    ): PropertiesPropertySource {
+        val factory =
+            YamlPropertiesFactoryBean().apply {
+                setResources(resource.resource)
+            }
+        return PropertiesPropertySource(resource.resource.filename ?: "unknown", factory.getObject()!!)
+    }
+}
