+++ hunmin-bootstrap/src/test/resources/application.yml
