+++ settings.gradle.kts
@@ -2,7 +2,7 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
 }
 rootProject.name = "hunmin-backend"
-include("hunmin-api")
+include("hunmin-bootstrap")
 include("hunmin-domain")
 include("hunmin-persistence")
 include("hunmin-common")
